+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Add CiaynAgent to support models that do not have, or are not good at, agentic function calling.
+- Improve env var validation.
 
 ## [0.10.3] - 2024-12-27